updated to use a decorator for BaseActor
